Finishing touches. Ready to be submit after review.
I included support for reading in only from standard in and outputting directly to the terminal in that case. I also cleaned up all errors and memory leaks that were outputting in the console. I also commented out all debug print messages. Give the program a quick run-through and double check everything, but it should be ready to submit.

Also feel free to add anything to the README before submission.